Don't show checkbox and delete button for default page.
git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/trunk@2942 3d254a34-79dc-0310-9e5f-be208747d8a0

+++ admin/listcontent.php
@@ -260,10 +260,17 @@ function display_hierarchy(&$root)
 			$thelist .= "<td class=\"pagepos\"><a href=\"editcontent.php?content_id=".$one->Id()."\">";
 			$thelist .= $editImg;;
 			$thelist .= "</a></td>\n";
+        if ($one->DefaultContent() != true) {
 			$thelist .= "<td class=\"pagepos\"><a href=\"deletecontent.php?content_id=".$one->Id()."\" onclick=\"return confirm('".lang('deleteconfirm')."');\">";
 			$thelist .= $deleteImg;
 			$thelist .= "</a></td>\n";
 			$thelist .= '<td class="pagepos"><input type="checkbox" name="multicontent-'.$one->Id().'" /></td>';
+		}
+		else
+		{
+			$thelist .= '<td>&nbsp;</td>' . "\n";
+			$thelist .= '<td>&nbsp;</td>' . "\n";
+		}
 			$thelist .= "</tr>\n";  	
 		}
 		else
